pyproject.toml
[tool.poetry]
name = "depot"
version = "0.1.0"
description = ""
authors = ["Ben Saltz <bensaltz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
flask = "^1.1.2"
requests-async = "^0.6.2"
flask_sqlalchemy = "^2.4.3"
flask-wtf = "^0.14.3"
python-dateutil = "^2.8.1"
flask-socketio = "^4.3.1"
eventlet = "^0.25.2"
flask-seasurf = "^0.2.2"
flask-migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
black = {extras = ["d"], version = "^19.10b0"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
